added rations, shared rations

+++ app/admin/rations.rb
@@ -1,7 +1,7 @@
 ActiveAdmin.register Ration do
   config.batch_actions = false
   #menu :priority => 1
-  scope_to :current_user
+  scope_to :current_user, :association_method => :own_rations
 
   filter :name
   filter :created_at
@@ -19,9 +19,11 @@ ActiveAdmin.register Ration do
     column :updated_at
 
     column "" do |resource|
-      links = ''.html_safe
-      links += link_to I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link"
-      links += link_to I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :data => {:confirm => I18n.t('active_admin.delete_confirmation')}, :class => "member_link delete_link"
+      links = "".html_safe
+      if current_user.id == resource.user.id
+        links += link_to I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link"
+        links += link_to I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :data => {:confirm => I18n.t('active_admin.delete_confirmation')}, :class => "member_link delete_link"
+      end
       links += current_user.setting(:ration).to_i == resource.id ? "Current " : link_to("Choose", admin_settings_update_path(:var => "ration", :value => resource.id, :ref => admin_rations_path), :method => :post, :class => "member_link")
 
       links
@@ -37,6 +39,15 @@ ActiveAdmin.register Ration do
 
     f.actions
   end
+
+  controller do
+    def index
+      index! do |format|
+        @rations = current_user.all_rations.page(params[:page])
+      end
+    end
+  end
+
 
 end
